well i messed up
I found a great open source model which has roughly 98% accuracy for guessing 6 digits captchas regardless of noise to a extent. But mind that the captchas I want to solve are Alphanumeric Case sensitive 6 digit captchas with a lot of wiggly lines (No IIT Delhi was harmed). Hence I would need to retrain this CNN Model from scratch. For that I would need to get character images which is manageable but also I would require the training data for which please go on the website to help me collect data to train